# docs/swagger.yaml
basePath: /
definitions:
...
  api.ClusterNode:
    properties:
      address:
        type: string
      core:
        $ref: '#/definitions/api.ClusterNodeCore'
      created_at:
        description: RFC 3339
        type: string
      error:
        type: string
      id:
        type: string
      last_contact_ms:
        type: number
      latency_ms:
        type: number
      leader:
        type: boolean
      name:
        type: string
      public_domains:
        items:
          type: string
        type: array
      resources:
        $ref: '#/definitions/api.ClusterNodeResources'
      status:
        type: string
      uptime_seconds:
        type: integer
      version:
        type: string
      voter:
        type: boolean
    type: object
  api.ClusterNodeCore:
    properties:
      address:
        type: string
      error:
        type: string
      last_contact_ms:
        description: milliseconds
        type: number
      latency_ms:
        description: milliseconds
        type: number
      status:
        type: string
      version:
        type: string
    type: object
  api.ClusterNodeFiles:
    properties:
      files:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      last_update:
        description: unix timestamp
        type: integer
    type: object
...
  api.ClusterNodeResources:
    properties:
      cpu_core:
        description: percent 0-100*ncpu
        type: number
      cpu_limit:
        description: percent 0-100*npcu
        type: number
      cpu_used:
        description: percent 0-100*npcu
        type: number
      error:
        type: string
      gpu:
        description: GPU resources
        items:
          $ref: '#/definitions/api.ClusterNodeGPUResources'
        type: array
      is_throttling:
        type: boolean
      memory_core_bytes:
        description: bytes
        type: integer
      memory_limit_bytes:
        description: bytes
        type: integer
      memory_total_bytes:
        description: bytes
        type: integer
      memory_used_bytes:
        description: bytes
        type: integer
      ncpu:
        type: number
    type: object
  api.ClusterNodeState:
    properties:
      state:
        enum:
        - online
        - maintenance
        - leave
        type: string
    required:
    - state
    type: object
...
  api.FileInfo:
    properties:
      core_id:
        type: string
      last_modified:
        format: int64
        type: integer
      name:
        type: string
      size_bytes:
        format: int64
        type: integer
    type: object
...
  api.FilesystemOperation:
    properties:
      bandwidth_limit_kbit:
        description: kbit/s
        type: integer
      operation:
        enum:
        - copy
        - move
        type: string
      source:
        type: string
      target:
        type: string
    required:
    - operation
    type: object
...
  api.Process:
    properties:
      config:
        $ref: '#/definitions/api.ProcessConfig'
      core_id:
        type: string
      created_at:
        format: int64
        type: integer
      domain:
        type: string
      id:
        type: string
      metadata: {}
      owner:
        type: string
      reference:
        type: string
      report:
        $ref: '#/definitions/api.ProcessReport'
      state:
        $ref: '#/definitions/api.ProcessState'
      type:
        type: string
      updated_at:
        format: int64
        type: integer
    type: object
  api.ProcessConfig:
    properties:
      autostart:
        type: boolean
      binary:
        type: string
      domain:
        type: string
      id:
        type: string
      input:
        items:
          $ref: '#/definitions/api.ProcessConfigIO'
        type: array
      limits:
        $ref: '#/definitions/api.ProcessConfigLimits'
      log_patterns:
        items:
          type: string
        type: array
      metadata:
        additionalProperties: true
        type: object
      options:
        items:
          type: string
        type: array
      output:
        items:
          $ref: '#/definitions/api.ProcessConfigIO'
        type: array
      owner:
        type: string
      reconnect:
        type: boolean
      reconnect_delay_seconds:
        format: uint64
        type: integer
      reference:
        type: string
      runtime_duration_seconds:
        format: uint64
        type: integer
      scheduler:
        type: string
      stale_timeout_seconds:
        format: uint64
        type: integer
      type:
        enum:
        - ffmpeg
        - ""
        type: string
    required:
    - input
    - output
    type: object
  api.ProcessConfigIO:
    properties:
      address:
        type: string
      cleanup:
        items:
          $ref: '#/definitions/api.ProcessConfigIOCleanup'
        type: array
      id:
        type: string
      options:
        items:
          type: string
        type: array
    required:
    - address
    type: object
...
  api.ProcessReport:
    properties:
      created_at:
        format: int64
        type: integer
      history:
        items:
          $ref: '#/definitions/api.ProcessReportHistoryEntry'
        type: array
      lines:
        additionalProperties:
          format: int64
          type: integer
        type: object
      log:
        items:
          items:
            type: string
          type: array
        type: array
      matches:
        items:
          type: string
        type: array
      prelude:
        items:
          type: string
        type: array
    type: object
...
  errors.Error:
    properties:
      code:
        format: int
        type: integer
      details:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
...
info:
  contact:
    email: hello@datarhei.com
    name: datarhei Core Support
    url: https://www.datarhei.com
  description: Expose REST API for the datarhei Core
  license:
    name: Apache 2.0
    url: https://github.com/datarhei/core/v16/blob/main/LICENSE
  title: datarhei Core API
  version: "3.0"
paths:
...
  /api/v3/cluster/node:
    get:
      description: List of proxy nodes in the cluster
      operationId: cluster-3-get-nodes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.ClusterNode'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: List of proxy nodes in the cluster
      tags:
      - v16.?.?
  /api/v3/cluster/node/{id}:
    get:
      description: List a proxy node by its ID
      operationId: cluster-3-get-node
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ClusterNode'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: List a proxy node by its ID
      tags:
      - v16.?.?
  /api/v3/cluster/node/{id}/files:
    get:
      description: List the resources of a proxy node by its ID
      operationId: cluster-3-get-node-files
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ClusterNodeFiles'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: List the resources of a proxy node by its ID
      tags:
      - v16.?.?
  /api/v3/cluster/node/{id}/fs:
    put:
      consumes:
      - application/json
      description: Execute file operations (copy or move) between registered filesystems
      operationId: cluster-3-node-fs-operation
      parameters:
      - description: Filesystem operation
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/api.FilesystemOperation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: File operations between filesystems
      tags:
      - v16.?.?
  /api/v3/cluster/node/{id}/fs/{storage}:
    get:
      description: List all files on a filesystem on a node. The listing can be ordered
        by name, size, or date of last modification in ascending or descending order.
      operationId: cluster-3-node-fs-list-files
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      - description: Name of the filesystem
        in: path
        name: storage
        required: true
        type: string
      - description: glob pattern for file names
        in: query
        name: glob
        type: string
      - description: none, name, size, lastmod
        in: query
        name: sort
        type: string
      - description: asc, desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.FileInfo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: List all files on a filesystem on a node
      tags:
      - v16.?.?
  /api/v3/cluster/node/{id}/fs/{storage}/{filepath}:
    delete:
      description: Remove a file from a filesystem on a node
      operationId: cluster-3-node-fs-delete-file
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      - description: Name of the filesystem
        in: path
        name: storage
        required: true
        type: string
      - description: Path to file
        in: path
        name: filepath
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Remove a file from a filesystem on a node
      tags:
      - v16.?.?
    get:
      description: Fetch a file from a filesystem on a node
      operationId: cluster-3-node-fs-get-file
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      - description: Name of the filesystem
        in: path
        name: storage
        required: true
        type: string
      - description: Path to file
        in: path
        name: filepath
        required: true
        type: string
      produces:
      - application/data
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "301":
          description: Moved Permanently
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Fetch a file from a filesystem on a node
      tags:
      - v16.?.?
    head:
      description: Fetch a file from a filesystem on a node
      operationId: cluster-3-node-fs-get-file
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      - description: Name of the filesystem
        in: path
        name: storage
        required: true
        type: string
      - description: Path to file
        in: path
        name: filepath
        required: true
        type: string
      produces:
      - application/data
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "301":
          description: Moved Permanently
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Fetch a file from a filesystem on a node
      tags:
      - v16.?.?
    put:
      consumes:
      - application/data
      description: Writes or overwrites a file on a filesystem on a node
      operationId: cluster-3-node-fs-put-file
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      - description: Name of the filesystem
        in: path
        name: storage
        required: true
        type: string
      - description: Path to file
        in: path
        name: filepath
        required: true
        type: string
      - description: File data
        in: body
        name: data
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - text/plain
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Add a file to a filesystem on a node
      tags:
      - v16.?.?
  /api/v3/cluster/node/{id}/process:
    get:
      description: List of processes in the cluster on a node
      operationId: cluster-3-list-node-processes
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      - description: Comma separated list of fields (config, state, report, metadata)
          that will be part of the output. If empty, all fields will be part of the
          output.
        in: query
        name: filter
        type: string
      - description: Return only these process that have this reference value. If
          empty, the reference will be ignored.
        in: query
        name: reference
        type: string
      - description: Comma separated list of process ids to list. Overrides the reference.
          If empty all IDs will be returned.
        in: query
        name: id
        type: string
      - description: Glob pattern for process IDs. If empty all IDs will be returned.
          Intersected with results from other pattern matches.
        in: query
        name: idpattern
        type: string
      - description: Glob pattern for process references. If empty all IDs will be
          returned. Intersected with results from other pattern matches.
        in: query
        name: refpattern
        type: string
      - description: Glob pattern for process owners. If empty all IDs will be returned.
          Intersected with results from other pattern matches.
        in: query
        name: ownerpattern
        type: string
      - description: Glob pattern for process domain. If empty all IDs will be returned.
          Intersected with results from other pattern matches.
        in: query
        name: domainpattern
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Process'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: List of processes in the cluster on a node
      tags:
      - v16.?.?
  /api/v3/cluster/node/{id}/state:
    get:
      description: Get the state of a node with the given ID
      operationId: cluster-3-get-node-state
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ClusterNodeState'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Get the state of a node with the given ID
      tags:
      - v16.?.?
    put:
      description: Set the state of a node with the given ID
      operationId: cluster-3-set-node-state
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      - description: State
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/api.ClusterNodeState'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Set the state of a node with the given ID
      tags:
      - v16.?.?
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  ApiRefreshKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  Auth0KeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"